#142 SynthV CFのzipでのみ発生するフォルダ名問題の発生条件判定誤りを修正。データ誤りを修正

+++ voice-list-generator/tuinavo/lib/Rename-SVCFFolders.ps1
@@ -1,14 +1,14 @@
 ﻿function Move-ChildFolders($zipNumber, $planName, $destDir) {
     foreach ($zn in $zipNumber1) {
-        Test-Path ("$destDir\Sample_voice_${_}_wav") {
+        if (Test-Path ("$destDir\Sample_voice_${zn}_wav\Sample_voice_${zn}_${pn}")) {
             Write-Host "SynthV CFリターンの第${_}回のフォルダ構成をFantiaで配布されているフォルダ構成に合わせます"
             foreach ($pn in $planName1.Keys ) {
                 Move-Item -Pass "$destDir\Sample_voice_${zn}_wav\Sample_voice_${zn}_${pn}" `
                     -Destination "$destDir\Sample_voice_${zn}_$($planName1[$pn])"
             } 
+            Remove-Item -Path "$destDir\Sample_voice_${zn}_wav"
+            Write-Host "第${zn}回のフォルダをリネームしました"
         }
-        Remove-Item -Path "$destDir\Sample_voice_${zn}_wav"
-        Write-Host "第${zn}回のフォルダをリネームしました"
     }
 }
 
@@ -30,7 +30,10 @@ function Rename-SVCFFolders($destDir) {
     Move-ChildFolders($zipNumber2, $planName2, $destDir)
 
     # XXX: 空フォルダ"wav_SynthesizerV CF向け限定ボイス（5000円プラン以上向け）"が作成されるので削除する
-    Test-Path ("$destDir\wav_SynthesizerV CF向け限定ボイス（5000円プラン以上向け）") {
-        Remove-Item -Path "$destDir\wav_SynthesizerV CF向け限定ボイス（5000円プラン以上向け）"
+    $svcf5000Folder = "$destDir\wav_SynthesizerV CF向け限定ボイス（5000円プラン以上向け）"
+    if (Test-Path ($svcf5000Folder)) {
+        if ((Get-ChildItem $svf5000Folder | Measure-Object ).count -eq 0) {
+            Remove-Item -Path "$destDir\wav_SynthesizerV CF向け限定ボイス（5000円プラン以上向け）"
+        }
     }
 }
